Provide better feedback when requests are sent
When a request is triggered there are three possible outcomes:

 - Request parameters are serialized, a valid response comes back and is decoded properly
 - Request parameters are serialized, a response comes back that can't be decoded
 - No request is sent because the state of the gui doesn't provide valid req params

Previously there was no way to tell when (3) was happening - request triggers were silently ignored.

To improve this situation, first we allow API parameters to report errors; captures become `Behavior t (Either String a)`, and QueryParam includes the ability to 'Either String' its parameter. Second, we augment the computed endpoints' return values to indicate the result of the request - one constructor for each of the possibilities listed above.