Use a nix builder instead of cabal-install
Caching was tried bit the travis cache mechanism is actually slower than
redownloading and rebuilding.

Added an alternative start to a cabal-based builder for compatibility
This requres the reflex reunification branch (or at least 18.09 from
some source.)

Note that all deps but reflex-platform are currently
unused, but the constraints also specify that at least the http-api-data
submodule is out of bounds for servant-reflex.